requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
cryptography==41.0.1
cssutils==2.7.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.7
django-allauth==0.58.2
django-braces==1.15.0
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.1.0
django-meta==2.4.0
django-mptt==0.15.0
django-mptt-admin==2.5.1
django-oauth-toolkit==2.3.0
django-redactor==0.2
django-redactoreditor==1.2.10
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
django-timezone-field==6.0.1
django-tinymce==3.6.1
djangorestframework==3.14.0
easy-thumbnails==2.8.5
google-auth==2.20.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
jwcrypto==1.5.0
kombu==5.3.0
lxml==4.9.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stem==1.8.2
Telethon==1.31.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
uWSGI==2.0.23
vine==5.0.0
wcwidth==0.2.6
wheel==0.41.2
whitenoise==3.3.1
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0